Use `int64_t`, not `long long`

+++ src/rep.c
@@ -1,5 +1,6 @@
 #include "vctrs.h"
 #include "utils.h"
+#include <inttypes.h> // For PRId64
 
 // Initialised at load time
 static struct vctrs_arg args_times_;
@@ -7,7 +8,7 @@ static struct vctrs_arg args_x_;
 static struct vctrs_arg* const args_times = &args_times_;
 static struct vctrs_arg* const args_x = &args_x_;
 
-static inline void stop_rep_size_oob(long long size);
+static inline void stop_rep_size_oob(int64_t size);
 static inline void stop_rep_times_size();
 static inline void stop_rep_times_negative();
 static inline void stop_rep_times_missing();
@@ -27,7 +28,7 @@ static SEXP vec_rep(SEXP x, R_len_t times) {
 
   const R_len_t x_size = vec_size(x);
 
-  const long long temp_size = (long long) x_size * times;
+  const int64_t temp_size = (int64_t) x_size * times;
   if (temp_size > INT_MAX) {
     stop_rep_size_oob(temp_size);
   }
@@ -80,7 +81,7 @@ static SEXP vec_rep_each_uniform(SEXP x, R_len_t times) {
 
   const R_len_t x_size = vec_size(x);
 
-  const long long temp_size = (long long) x_size * times;
+  const int64_t temp_size = (int64_t) x_size * times;
   if (temp_size > INT_MAX) {
     stop_rep_size_oob(temp_size);
   }
@@ -113,7 +114,7 @@ static SEXP vec_rep_each_impl(SEXP x, SEXP times, const R_len_t times_size) {
 
   const int* p_times = INTEGER_RO(times);
 
-  long long temp_size = 0;
+  int64_t temp_size = 0;
   for (R_len_t i = 0; i < times_size; ++i) {
     const int elt_times = p_times[i];
 
@@ -178,8 +179,8 @@ SEXP vctrs_rep_each(SEXP x, SEXP times) {
 
 // -----------------------------------------------------------------------------
 
-static inline void stop_rep_size_oob(long long size) {
-  Rf_errorcall(R_NilValue, "Requested output size must be less than %i, not %lli.", INT_MAX, size);
+static inline void stop_rep_size_oob(int64_t size) {
+  Rf_errorcall(R_NilValue, "Requested output size must be less than %i, not %" PRId64 ".", INT_MAX, size);
 }
 
 static inline void stop_rep_times_size() {
